Improve aggregates Work on test_rsa
